[REFACTOR] #226 - VideoCaptureView 리팩토링
- 촬영 버튼이 있던 recordingSection을 삭제하고, recordingButton 하나의 viewbuilder로 대체하였습니다.
- 촬영 버튼을 눌렀을 때의 애니메이션을 추가하고, 컴포넌트의 간격과 패딩값을 맞췄습니다.
- enum을 사용해 촬영 상태에 따른 버튼의 크기 및 형태를 조절하였습니다.
- 기존의 enum buttonState의 이름을 VideoCaptureState로 변경하고, 각 상태에 따른 뷰의 속성값을 추가했습니다.
- 한 파일에 몰려있던 viewbuilder, func들을 별도의 파일로 분리했습니다. VideoCaptureView+ViewBuilder.swift 와 VideoCaptureView+Func.swift입니다.
