+++ Whistle/UI/Screens/VideoEditor/VideoEditorView.swift
@@ -107,6 +107,10 @@ struct VideoEditorView: View {
           }
         }
       }
+      .gesture(DragGesture().onEnded { _ in
+
+      })
+      .disableSwipeBack()
       .toolbar(.hidden)
       .background(Color.Background_Default_Dark)
       .ignoresSafeArea()
